% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerFlux.R
\docType{data}
\name{df_dndc}
\alias{df_dndc}
\title{DNDC Model output data}
\format{
## `df_dndc`
A data frame with 88 rows and 5 columns:
\describe{
  \item{site}{Site name}
  \item{days_since_fert, secs_since_fert}{Time since fertiliser applied in days or seconds}
  \item{Fn2o_kgN_ha_d, Fn2o_nmol_m2_s}{Predicted flux of N2O in given units}
}
}
\source{
<https://www.dndc.sr.unh.edu/>
}
\usage{
df_dndc
}
\description{
Output data from some example runs of the DNDC model
}
\keyword{datasets}
