% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerFlux.R
\name{get_ci_flux}
\alias{get_ci_flux}
\title{get_ci_flux}
\usage{
get_ci_flux(
  sigma_chi = set_units(25, nmol_n2o/mol),
  height = set_units(0.23, m),
  sigma_t = set_units(90, s),
  n_gas = 10
)
}
\arguments{
\item{sigma_chi}{PARAM_DESCRIPTION, Default: set_units(25, nmol_n2o/mol)}

\item{height}{PARAM_DESCRIPTION, Default: set_units(0.23, m)}

\item{sigma_t}{PARAM_DESCRIPTION, Default: set_units(90, s)}

\item{n_gas}{PARAM_DESCRIPTION, Default: 10}
}
\value{
The 95 % confidence interval for a flux measurement
}
\description{
Calculates the 95 % confidence interval for a flux measurement
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 # sigma_t <- sapply(n_gas, get_sigma_t, x_max = t_max)
 }
}
}
