% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerFlux.R
\name{get_sigma_spatial}
\alias{get_sigma_spatial}
\title{This is an attempt to compare sampling from normal versus lognormal
distributions. Not sure we can make these comparable meaningfully.}
\usage{
get_sigma_spatial(
  n_samples = 10,
  n_sims = 1e+05,
  location = 0,
  scale = 1.5,
  dist_lognormal = TRUE
)
}
\description{
This is an attempt to compare sampling from normal versus lognormal
distributions. Not sure we can make these comparable meaningfully.
}
