% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerFlux.R
\name{get_sigma_t}
\alias{get_sigma_t}
\title{get_sigma_t}
\usage{
get_sigma_t(v_t)
}
\arguments{
\item{v_t}{A vector of times}
}
\value{
the standard deviation
}
\description{
Get the standard deviation of the time series
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
