% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerFlux.R
\name{get_v_t}
\alias{get_v_t}
\title{get_v_t}
\usage{
get_v_t(x_max, n_gas)
}
\arguments{
\item{x_max}{Maximum time for chamber enclosure}

\item{n_gas}{Number of gas samples taken over the measurement}
}
\value{
A vector of evenly spaced times
}
\description{
Get a vector of n_gas evenly spaced times up to x_max
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
