% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerFlux-package.R
\docType{package}
\name{powerFlux-package}
\alias{powerFlux}
\alias{powerFlux-package}
\title{powerFlux: Power Flux}
\description{
This package is provides code for statistical power analysis of experiments based on chamber flux measurements, particularly those looking at emission factors for N2O.
}
\author{
\strong{Maintainer}: Peter Levy \email{plevy@ceh.ac.uk} (\href{https://orcid.org/0000-0002-8505-1901}{ORCID}) [copyright holder]

}
\keyword{internal}
