% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerFlux.R
\name{rmse}
\alias{rmse}
\title{RMSE}
\usage{
rmse(error)
}
\arguments{
\item{error}{Vector of differences betwwen predicted and observed values}
}
\value{
Root Mean Squared Error
}
\description{
Function that returns Root Mean Squared Error
}
\details{
DETAILS
}
